+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 ### Fixes
 
 - Fixed conflicting default name for scriptable options configuration scripts ([#1146](https://github.com/getsentry/sentry-unity/pull/1146))
+- Made inlined helpers on the macOS bridge static ([#1148](https://github.com/getsentry/sentry-unity/pull/1148))
 
 ### Dependencies